Convert tests to use ReactDOM
